fix-issue-803(1) Remove _FPref32 suffix from verification cache files & factor out related code. Compute reference data for convolutions as floats (FP32). Aside tidy fix.

+++ driver/main.cpp
@@ -50,7 +50,7 @@ int main(int argc, char* argv[])
     Driver* drv;
     if(base_arg == "conv")
     {
-        drv = new ConvDriver<float, double>();
+        drv = new ConvDriver<float, float>();
     }
     else if(base_arg == "convfp16")
     {
